requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.5
pydantic>=2.0.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4
langchain>=0.0.267
openai>=0.27.0
sentence-transformers>=2.2.2
